#853 Separate the S3 network fetcher from range-cache dispatch
S3InputFile held both the signed-range HTTP logic and the decision
whether a read should be served from the SPARSE_TEMPFILE cache. That
forced an `if (cache != null)` branch into open() and readRange(), and
required a BareFetcher inner class purely so the cache could call back
into the network path without recursing through readRange() and
short-circuiting to itself again.

Splitting the network path out into a package-private S3Fetcher lets
RangeBackedInputFile wrap it the way a decorator is meant to, with the
fetcher unaware that a cache exists. S3InputFile stays the public,
documented return type of S3Source.inputFile(...) — it becomes a facade
that picks its delegate once in the constructor, so no call site and no
counter semantics change. Inverting the composition instead would have
restored the shape #560 reverted.

length() now declares the IOException that InputFile#length permits.
Omitting it constrained every implementation reachable through the
facade to determine the length without I/O, which is an accident of how
S3InputFile happened to be written rather than something the contract
guarantees. No caller in the repository needed adjusting — they all go
through InputFile, which already declares it.

RangeBackedInputFile's delegate() and isPopulated() go with it. Both
were declared for test and diagnostic use and neither ever acquired a
caller; the one construct that motivated an accessor for the wrapped
file is gone, and the cache's behaviour is asserted through the network
counters rather than by inspecting its populated set.

The temp directory is validated where it is configured rather than at
the first open(): a caller who points tempDir at a path that does not
exist learns about it while building the S3Source, not several frames
into a read. Only SPARSE_TEMPFILE consults it, so NONE keeps accepting
anything.

close() no longer fails when the platform refuses to unlink a file that
still carries a mapping. Windows does refuse, the unmap here is
GC-driven, and dive opts every S3 file into the range cache — so what
would be modelled as an exceptional delete failure is the ordinary case
on a platform the CLI ships binaries for. The delete falls back to JVM
exit and logs. Making cleanup non-throwing also stops it masking the
original failure on the open() error path, where the cleanup ran inside
the catch block that was about to rethrow.
