+++ s3/src/main/java/dev/hardwood/s3/S3InputFile.java
@@ -8,26 +8,24 @@
 package dev.hardwood.s3;
 
 import java.io.IOException;
-import java.io.InputStream;
-import java.net.http.HttpResponse;
 import java.nio.ByteBuffer;
-import java.nio.file.Path;
-import java.util.concurrent.atomic.AtomicLong;
 
 import dev.hardwood.InputFile;
-import dev.hardwood.internal.FetchReason;
 import dev.hardwood.internal.reader.RangeBackedInputFile;
-import dev.hardwood.s3.internal.S3Api;
 
 /// [InputFile] backed by an object in Amazon S3 (or an S3-compatible service).
 ///
-/// Each [#readRange] call issues a signed HTTP `GET` request with a
-/// byte-range header, so only the requested bytes are transferred.
+/// A [#readRange] call that cannot be answered locally issues a signed
+/// HTTP `GET` request with a byte-range header, so only the requested
+/// bytes are transferred.
 ///
 /// [#open()] uses a suffix-range GET instead of a HEAD request. This
 /// discovers the file length from the `Content-Range` response header
 /// and pre-fetches the Parquet footer (which sits at the end of the file) in
 /// the same round-trip — eliminating a separate HEAD request per file.
+/// Those tail bytes are retained for as long as the file is open, and a
+/// read falling entirely inside that window is answered from them
+/// without a request.
 ///
 /// When the owning [S3Source] is configured with
 /// [RangeBacking#SPARSE_TEMPFILE], an internal mmap-backed range cache
@@ -39,224 +37,61 @@ import dev.hardwood.s3.internal.S3Api;
 /// Thread-safe once [#open()] has been called.
 public class S3InputFile implements InputFile {
 
-    private static final System.Logger LOG = System.getLogger(S3InputFile.class.getName());
+    /// The S3 network path, retained so the counters report network
+    /// traffic only: under [RangeBacking#SPARSE_TEMPFILE] a cache hit is
+    /// served from the mapping and never reaches the fetcher.
+    private final S3Fetcher fetcher;
 
-    /// Number of bytes to fetch from the tail of the file during [#open()].
-    /// 64 KB is large enough to cover the Parquet footer in virtually all files
-    /// (footer + footer length + magic is typically a few KB).
-    static final int TAIL_SIZE = 64 * 1024;
-
-    private final S3Api api;
-    private final String bucket;
-    private final String key;
-    private long fileLength = -1;
-    private ByteBuffer tailCache;
-    private long tailCacheOffset;
-    private final AtomicLong networkRequestCount = new AtomicLong();
-    private final AtomicLong networkBytesFetched = new AtomicLong();
-    /// Non-null iff the owning [S3Source] was configured with
-    /// [RangeBacking#SPARSE_TEMPFILE]. Wraps a private adapter that
-    /// exposes the bare S3 fetch path, so cache hits don't re-enter
-    /// [#readRange] (and thus don't double-count network counters).
-    private final RangeBackedInputFile cache;
+    /// What the [InputFile] methods delegate to: either [#fetcher]
+    /// itself, or the range cache wrapping it. Chosen once at
+    /// construction, so the read path carries no per-call dispatch.
+    private final InputFile impl;
 
     S3InputFile(S3Source source, String bucket, String key) {
-        this.api = source.api();
-        this.bucket = bucket;
-        this.key = key;
-        Path tempDir = source.rangeBacking() == RangeBacking.SPARSE_TEMPFILE
-                ? source.tempDir()
-                : null;
-        this.cache = tempDir != null ? new RangeBackedInputFile(new BareFetcher(), tempDir) : null;
+        this.fetcher = new S3Fetcher(source.api(), bucket, key);
+        this.impl = source.rangeBacking() == RangeBacking.SPARSE_TEMPFILE
+                ? new RangeBackedInputFile(fetcher, source.tempDir())
+                : fetcher;
     }
 
     @Override
     public void open() throws IOException {
-        if (cache != null) {
-            cache.open();
-            return;
-        }
-        openBare();
+        impl.open();
     }
 
     @Override
     public ByteBuffer readRange(long offset, int length) throws IOException {
-        if (cache != null) {
-            return cache.readRange(offset, length);
-        }
-        return readRangeBare(offset, length);
-    }
-
-    private void openBare() throws IOException {
-        if (fileLength >= 0) {
-            return;
-        }
-        String suffixRange = "bytes=-" + TAIL_SIZE;
-        HttpResponse<byte[]> response = api.getBytes(bucket, key, suffixRange);
-        int status = response.statusCode();
-        if (status != 206 && status != 200) {
-            throw new IOException("Failed to open " + name()
-                    + ": HTTP " + status + " " + new String(response.body()));
-        }
-        fileLength = parseFileLength(response);
-        byte[] tail = response.body();
-        long requestNo = networkRequestCount.incrementAndGet();
-        long totalBytes = networkBytesFetched.addAndGet(tail.length);
-        // Use a direct buffer so slices are usable from FFM-based decompressors
-        // (e.g. libdeflate), which require native MemorySegments.
-        tailCache = ByteBuffer.allocateDirect(tail.length);
-        tailCache.put(tail);
-        tailCache.flip();
-        tailCacheOffset = fileLength - tail.length;
-        logFetch("open-tail", tailCacheOffset, tail.length, requestNo, totalBytes);
-    }
-
-    private ByteBuffer readRangeBare(long offset, int length) throws IOException {
-        // Serve from the tail cache if the requested range falls within it
-        if (tailCache != null && offset >= tailCacheOffset
-                && offset + length <= tailCacheOffset + tailCache.capacity()) {
-            int relOffset = Math.toIntExact(offset - tailCacheOffset);
-            LOG.log(System.Logger.Level.DEBUG,
-                    "[{0}] readRange offset={1} length={2} reason={3} (tail cache hit)",
-                    name(), offset, length, FetchReason.current());
-            return tailCache.slice(relOffset, length);
-        }
-
-        long requestNo = networkRequestCount.incrementAndGet();
-        long totalBytes = networkBytesFetched.addAndGet(length);
-        logFetch(FetchReason.current(), offset, length, requestNo, totalBytes);
-        String range = "bytes=" + offset + "-" + (offset + length - 1);
-        HttpResponse<InputStream> response = api.getStream(bucket, key, range);
-        int status = response.statusCode();
-        if (status != 206 && status != 200) {
-            try (InputStream body = response.body()) {
-                throw new IOException("Failed to read range [" + offset + ", " + (offset + length)
-                        + ") from " + name() + ": HTTP " + status + " " + new String(body.readAllBytes()));
-            }
-        }
-        try (InputStream stream = response.body()) {
-            ByteBuffer buf = ByteBuffer.allocateDirect(length);
-            byte[] tmp = new byte[Math.min(8192, length)];
-            while (buf.hasRemaining()) {
-                int toRead = Math.min(tmp.length, buf.remaining());
-                int read = stream.read(tmp, 0, toRead);
-                if (read < 0) {
-                    break;
-                }
-                buf.put(tmp, 0, read);
-            }
-            if (buf.position() != length) {
-                throw new IOException("Short read from " + name() + ": expected " + length
-                        + " bytes but received " + buf.position());
-            }
-            buf.flip();
-            return buf;
-        }
+        return impl.readRange(offset, length);
     }
 
     @Override
-    public long length() {
-        if (fileLength < 0) {
-            throw new IllegalStateException("File not opened: " + name());
-        }
-        return fileLength;
+    public long length() throws IOException {
+        return impl.length();
     }
 
     @Override
     public String name() {
-        return "s3://" + bucket + "/" + key;
+        return impl.name();
     }
 
     @Override
     public void close() throws IOException {
-        if (cache != null) {
-            cache.close();
-        }
+        impl.close();
     }
 
     /// Number of HTTP requests issued against the object since [#open()].
     /// Counts the suffix-range tail fetch from `open` plus every
-    /// network-fetch [#readRange] call. Tail-cache hits do not count.
+    /// network-fetch [#readRange] call. Tail-cache and range-cache hits
+    /// do not count.
     public long networkRequestCount() {
-        return networkRequestCount.get();
+        return fetcher.networkRequestCount();
     }
 
     /// Number of bytes fetched from the network since [#open()]. The tail
     /// fetch from `open` contributes its actual response size; each
     /// network-fetch [#readRange] contributes the requested length.
-    /// Tail-cache hits do not count.
+    /// Tail-cache and range-cache hits do not count.
     public long networkBytesFetched() {
-        return networkBytesFetched.get();
-    }
-
-    /// Emits a single FINE log line per network fetch carrying the byte
-    /// range, the [FetchReason] tag attributed to the caller, and the
-    /// running per-file totals. Designed to be greppable: one line per
-    /// fetch, no nested structure.
-    ///
-    /// Logged at `System.Logger.Level.DEBUG`, which the platform logger
-    /// finder maps to JUL `Level.FINE` — `DiveCommand` configures the
-    /// JUL handler at `Level.FINE`, so a refactor that lowers the level
-    /// here would silently drop dive's `--log-file` output.
-    private void logFetch(String reason, long offset, int length, long requestNo, long totalBytes) {
-        LOG.log(System.Logger.Level.DEBUG,
-                "[{0}] fetch #{1} reason={2} offset={3} length={4} range=[{3},{5}) totalBytes={6}",
-                name(), requestNo, reason, offset, length, offset + length, totalBytes);
-    }
-
-    /// Extracts the total file length from the HTTP response.
-    ///
-    /// For suffix-range requests, S3 returns a `Content-Range` header like
-    /// `bytes 1000-1999/2000` where the number after `/` is the total
-    /// object size. If the header is absent (e.g. file smaller than TAIL_SIZE),
-    /// falls back to `Content-Length`.
-    private static long parseFileLength(HttpResponse<?> response) throws IOException {
-        String contentRange = response.headers().firstValue("Content-Range").orElse(null);
-        if (contentRange != null) {
-            int slashIdx = contentRange.lastIndexOf('/');
-            if (slashIdx >= 0) {
-                try {
-                    return Long.parseLong(contentRange.substring(slashIdx + 1));
-                }
-                catch (NumberFormatException e) {
-                    throw new IOException("Failed to parse Content-Range header: " + contentRange, e);
-                }
-            }
-        }
-        return response.headers().firstValueAsLong("Content-Length")
-                .orElseThrow(() -> new IOException("Response missing both Content-Range and Content-Length headers"));
-    }
-
-    /// Bare-fetch adapter handed to the [RangeBackedInputFile] cache so
-    /// the cache delegates back into the enclosing [S3InputFile]'s
-    /// network path *without* recursing through [#readRange] (which
-    /// would short-circuit to the cache again). The adapter's
-    /// [#close()] is a no-op because the enclosing instance owns no
-    /// per-file resources to release.
-    private final class BareFetcher implements InputFile {
-
-        @Override
-        public void open() throws IOException {
-            openBare();
-        }
-
-        @Override
-        public ByteBuffer readRange(long offset, int length) throws IOException {
-            return readRangeBare(offset, length);
-        }
-
-        @Override
-        public long length() {
-            return S3InputFile.this.length();
-        }
-
-        @Override
-        public String name() {
-            return S3InputFile.this.name();
-        }
-
-        @Override
-        public void close() {
-        }
+        return fetcher.networkBytesFetched();
     }
 }
